+++ src/crypto/CryptoNight_x86.h
@@ -484,8 +484,8 @@ inline void cryptonight_single_hash(const uint8_t *__restrict__ input, size_t si
 
             ((int64_t*)&l0[idx0 & MASK])[0] = n ^ q;
 
-            if (VARIANT == xmrig::VARIANT_XHV) {
-                idx0 = (~d) ^ q;
+            if (VARIANT == xmrig::VARIANT_ITA) {
+                idx0 = ~(d ^ q);
             }
             else {
                 idx0 = d ^ q;
@@ -599,8 +599,8 @@ inline void cryptonight_double_hash(const uint8_t *__restrict__ input, size_t si
 
             ((int64_t*)&l0[idx0 & MASK])[0] = n ^ q;
 
-            if (VARIANT == xmrig::VARIANT_XHV) {
-                idx0 = (~d) ^ q;
+            if (VARIANT == xmrig::VARIANT_ITA) {
+                idx0 = ~(d ^ q);
             }
             else {
                 idx0 = d ^ q;
@@ -639,8 +639,8 @@ inline void cryptonight_double_hash(const uint8_t *__restrict__ input, size_t si
 
             ((int64_t*)&l1[idx1 & MASK])[0] = n ^ q;
 
-            if (VARIANT == xmrig::VARIANT_XHV) {
-                idx1 = (~d) ^ q;
+            if (VARIANT == xmrig::VARIANT_ITA) {
+                idx1 = ~(d ^ q);
             }
             else {
                 idx1 = d ^ q;
@@ -708,8 +708,8 @@ inline void cryptonight_double_hash(const uint8_t *__restrict__ input, size_t si
         int32_t d = ((int32_t*)&l[idx & MASK])[2];      \
         int64_t q = n / (d | 0x5);                      \
         ((int64_t*)&l[idx & MASK])[0] = n ^ q;          \
-        if (VARIANT == xmrig::VARIANT_XHV) {            \
-            idx = (~d) ^ q;                            \
+        if (VARIANT == xmrig::VARIANT_ITA) {            \
+            idx = ~(d ^ q);                            \
         }                                               \
         else {                                          \
             idx = d ^ q;                               \